Move  IInsightViewProps, IInsightTitleProps and code generator to internal
