+++ libs/sdk-ui-ext/src/insightView/index.ts
@@ -1,9 +1,9 @@
-// (C) 2020-2021 GoodData Corporation
+// (C) 2020-2022 GoodData Corporation
 
 export { InsightView } from "./InsightView";
-export { IInsightViewProps } from "./types";
-export { IInsightTitleProps } from "./InsightTitle";
 
 // exports only for internal use by other gooddata packages
 export { IInsightErrorProps, InsightError } from "./InsightError";
 export { IInsightRendererProps, InsightRenderer } from "./InsightRenderer";
+
+export { IInsightViewProps, IInsightTitleProps } from "../internal/index";
