+++ libs/sdk-ui-ext/src/internal/utils/embeddingInsightViewCodeGenerator/insightViewCodeGenUtils.ts
@@ -1,7 +1,7 @@
 // (C) 2022 GoodData Corporation
 
 import { isIdentifierRef, isUriRef } from "@gooddata/sdk-model";
-import { IAdditionalFactoryDefinition } from "../../internal/utils/embeddingCodeGenerator/types";
+import { IAdditionalFactoryDefinition } from "../embeddingCodeGenerator/types";
 
 interface IInsightViewAdditionTransformationOptions {
     uriRef: string;
