+++ libs/sdk-ui-ext/src/internal/utils/embeddingInsightViewCodeGenerator/tests/insightViewCodeGenerator.test.ts
